Refactor loading spinner in AppButton and update top bar background color
The loading spinner in the AppButton component has been refactored to improve its animation. Instead of directly rendering the spinner icon, it is now wrapped in a div with the "animate-spin" class. This change enhances the visual experience for users when the button is in a loading state.

Additionally, the background color of the top bar in the AppTopBar component has been updated. The previous background color class "bg-light-page-bg" has been replaced with "bg-light-page-bg2" for better visual consistency.