Refactor AppCodeInput component to support customizable box dimensions
The AppCodeInput component has been refactored to allow for customization of the box dimensions. This change introduces two new optional props, `boxWidth` and `boxHeight`, which can be used to specify the width and height of the input boxes respectively. This enhancement provides more flexibility in adapting the component to different design requirements.